+++ crates/wdl-lint/src/rules/lint_directive_formatted.rs
@@ -88,7 +88,7 @@ impl Rule for LintDirectiveFormattedRule {
     }
 
     fn tags(&self) -> TagSet {
-        TagSet::new(&[Tag::Clarity, Tag::Correctness])
+        TagSet::new(&[Tag::Clarity, Tag::Correctness, Tag::SprocketCompatibility])
     }
 
     fn exceptable_nodes(&self) -> Option<&'static [wdl_ast::SyntaxKind]> {