+++ crates/wdl-lint/src/rules/matching_output_meta.rs
@@ -125,7 +125,11 @@ impl Rule for MatchingOutputMetaRule<'_> {
     }
 
     fn tags(&self) -> TagSet {
-        TagSet::new(&[Tag::Completeness, Tag::Documentation])
+        TagSet::new(&[
+            Tag::Completeness,
+            Tag::Documentation,
+            Tag::SprocketCompatibility,
+        ])
     }
 
     fn exceptable_nodes(&self) -> Option<&'static [SyntaxKind]> {
@@ -143,6 +147,7 @@ impl Rule for MatchingOutputMetaRule<'_> {
             "OutputSection",
             "RequirementsSection",
             "RuntimeSection",
+            "DescriptionLength",
         ]
     }
 }