+++ crates/wdl-lint/src/rules/meta_sections.rs
@@ -132,6 +132,7 @@ impl Rule for MetaSectionsRule {
             "RequirementsSection",
             "RuntimeSection",
             "MatchingOutputMeta",
+            "DescriptionLength",
         ]
     }
 }