+++ crates/wdl-lint/src/rules/parameter_meta_matched.rs
@@ -115,7 +115,12 @@ impl Rule for ParameterMetaMatchedRule {
     }
 
     fn tags(&self) -> TagSet {
-        TagSet::new(&[Tag::Completeness, Tag::Sorting, Tag::Documentation])
+        TagSet::new(&[
+            Tag::Completeness,
+            Tag::Sorting,
+            Tag::Documentation,
+            Tag::SprocketCompatibility,
+        ])
     }
 
     fn exceptable_nodes(&self) -> Option<&'static [SyntaxKind]> {
@@ -133,6 +138,7 @@ impl Rule for ParameterMetaMatchedRule {
             "RequirementsSection",
             "RuntimeSection",
             "MatchingOutputMeta",
+            "DescriptionLength",
         ]
     }
 }