+++ crates/wdl-lint/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* New `Tag::SprocketCompatibility` ([#351](https://github.com/stjude-rust-labs/sprocket/pull/351)).
+* New lint rule `DescriptionLength` ([#351](https://github.com/stjude-rust-labs/sprocket/pull/351)).
+
 ## 0.16.0 - 09-15-2025
 
 #### Added